pyproject.toml
[tool.poetry]
name = "pdf-search"
version = "0.1.0"
description = ""
authors = ["fruit-bird <github.ru3w9@slmail.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.32"
fastapi = "^0.112.2"
uvicorn = "^0.30.6"
pydantic-settings = "^2.4.0"
alembic = "^1.13.2"
pypdf = "^4.3.1"
pyyaml = "^6.0.2"
langchain = "^0.2.15"
langchain-community = "^0.2.14"
python-multipart = "^0.0.9"
tiktoken = "^0.7.0"
langchain-openai = "^0.1.23"
psycopg2-binary = "^2.9.9"
asyncio = "^3.4.3"
aiofiles = "^24.1.0"
aioboto3 = "^13.1.1"
sentence-transformers = "^3.0.1"
pdfminer-six = "^20240706"
chromadb = "^0.5.5"

[tool.poetry.group.test.dependencies]
requests = "^2.32.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
